LPD-15494 Remove useless abstraction
